// demo-mcp-21-code-generator-node/input/reference-repos/MSPrChValGenerate/src/common/utils/enums/taks.enum.ts
export enum ETaskDesc {
    DOWNLOAD_FILE = "ERROR EN LA DESCARGA POR MEDIO DEL SFTP",
    GET_PARAMS = 'ERROR CONSULTANDO PARAMS',
    TRANSFORMDATA = "ERROR DESCARGA Y TRANSFORMACION DE DATA ",
    READ_JSON = "ERROR EN LA TRANSFORMACION DE LINEA DEL JSON DE FEATURES",
    FIND_CARACTERISTICAS='ERROR CONSULTANDO CARACTERISTICAS CON DISPONIBILIDAD',
    SAVE_DATA='ERROR GUARDANDO DATA',
    GET_ALL_TERMS_CONDITIONS='ERROR OBTENIENDO TODA LA DATA',
    CONEXION_DB_SP = "ERROR DE CONEXION A LA BASE DE DATOS",
    TRANSFORM_DATA = "ERROR EN LA TRANSFORMACION DE DATOS",
    WRITE_FILE='ERROR ESCRIBIENDO ARCHIVO VOLCADO',
    FIND_DATE= "ERROR BUSCANDO FECHA ARCHIVO",
    EXCEPTION_SKU = "ERROR CRUCE EXCEPCIONES SKU",
    DELETE_DATA = "ERRRO ELIMINANDO DATA",
    ERROR_GET_DATA_DISPONIBILITY_FILE = 'ERROR_GET_DATA_DISPONIBILITY_FILE',
    ERROR_EXECUTE_REST = 'ERROR_SERVICE_OMN',
    REMOVE = 'ERROR ELIMINANDO ARCHIVO',
    ERROR_FILEDATA = 'ERROR LEYENDO LA DATA DEL ARCHIVO',
    ERROR_DATA = 'ERROR EN LA DATA',
    ERROR_TRANSFORM_PRICES_EQU_TECH = 'ERROR EN LA TRANSFORMACION DE PRECIOS EN TERMINALES Y TECNOLOGIA',


}